chore(platformtools): split slack tools per file
Address review feedback: one `tool_*.go` file per Slack tool, rename
`logs/search_logs.go` to follow the same convention, and simplify the
changeset for customer-facing release notes.
