+++ server/internal/platformtools/logs/tool_search_logs.go
